added user check for edit and deleting campgrounds

+++ routes/campgroundRoutes.js
@@ -63,6 +63,11 @@ router.put(
 	isLoggedIn,
 	catchAsync(async (req, res) => {
 		const id = req.params.id;
+		const camp = await Campground.findById(id);
+		if (!camp.author.equals(req.user._id)) {
+			req.flash('error', 'You do not have permission to do this!');
+			return res.redirect(`/campgrounds/${id}`);
+		}
 		const campground = await Campground.findByIdAndUpdate(id, {
 			...req.body.campground
 		});
@@ -76,10 +81,14 @@ router.get(
 	isLoggedIn,
 	catchAsync(async (req, res) => {
 		const id = req.params.id;
-		const campground = await Campground.findById(id);
+		const camp = await Campground.findById(id);
 		if (!campground) {
 			req.flash('error', 'Campground does not exist!');
 			return res.redirect('/campgrounds');
+		}
+		if (!camp.author.equals(req.user._id)) {
+			req.flash('error', 'You do not have permission to do this!');
+			return res.redirect(`/campgrounds/${id}`);
 		}
 		res.render('campgrounds/edit', { campground });
 	})
@@ -90,6 +99,11 @@ router.delete(
 	isLoggedIn,
 	catchAsync(async (req, res) => {
 		const id = req.params.id;
+		const camp = await Campground.findById(id);
+		if (!camp.author.equals(req.user._id)) {
+			req.flash('error', 'You do not have permission to do this!');
+			return res.redirect(`/campgrounds/${id}`);
+		}
 		const campground = await Campground.findByIdAndDelete(id);
 		req.flash('success', 'Successfully deleted campground!');
 		res.redirect('/campgrounds');
